after create CoinPayPlan  and change relation of payment

+++ CMS_Golbarg/Areas/Admin/Models/Mixer.cs
@@ -44,7 +44,7 @@ namespace CMS_Golbarg.Areas.Admin.Models
         [ForeignKey("PaintingWayId")]
         public PaintingWay PaintingWay { set; get; }
 
-        public byte PaintingWayId { set; get; }
+        public int PaintingWayId { set; get; }
 
     }
 }
